requirements.txt
django>=5.0
djangorestframework
django-cors-headers
channels
daphne
opencv-python-headless
ultralytics
torch
torchvision
numpy
dlib
scipy
imutils
werkzeug
pillow
huggingface-hub
psutil
openpyxl
mediapipe>=0.10.30
